refactor(connectors): extract resolveStringReference to shared module
